Medium Commit [Pre-Alpha v1] [Snapshot 0.3]
Made code more manageable.

Largest introduction is Networking. Base networking is in place, the
server can be run as a build and that build has authority over all game
objects aside from the Client player object and Client.cs script that
has local player authority on whichever client launches it.

Need to figure out [Command] functions and how to make the server owned
unit move when a client owned unit is moved locally so that it updates
for everyone.