Base Unit class and various UI.
Create a base unit class from which I intend to inherit onto all units.
It will activate a reference UI menu, comes with a Move() function that
takes a destination and will move the unit towards that destination by a
given time (temporarily deltaTime * movementSpeed).  The unit class has
a parent nation and has an access matrix that, at the moment only
controls interaction and visiblity. Access matrix is set based on parent
nation and the clients nation. At the moment, if the client's nation is
not the same as the unit's, the unit will not show on the map and will
not be clickable.

Added some temporary UI to control the unit.