Camera Control fully implemented.
Full implemented camera controls. When the game is launched, the camera
will pan to whichever nation is chosen. Positional values are pretty
accurate at the moment, but made it modular so that it can be changed
easily in the future.

The camera can be dragged around with the right mouse button, the camera
will move opposite to the mouse (like mobile scrolling). The camera will
be pushed back slightly if it goes out of bounds.

The camera can be zoomed in and out, changing orthographicSize up and
down. Limited to > 2 < 50 size.

Minor code changes to the GameManager.cs, just to help with script -
script communication.