Version 1 Infantry Unit.
Created a version 1 infantry unit that inherits from the base Unit
class. Changed up some inheritance stuff. On the base Unit class, the
monobehaviour functions Start() and Update() are removed and only the
inheritinng class should call those functions. The base Unit class
contains a unitStart() and a unitUpdate(), this are both public and get
called on the inheriting class via their Start() and Update() functions,
essentially emulating the monobehaviour functions on the base Unit
class, without restricting inheriting classes from using Start() and
Update() without breaking the base Unit class setup and update.

Rudimentary movement implemented, when StartMovement() is called, a
destination is passed in and isMoving is set to true. In the
unitUpdate() function, if isMoving is true, Move() is called which will
move the unit towards the destination at the movementSpeed * deltaTime.
Values are all rudimentary, delta time may not be used as the game will
not be frame dependant due to its realtime nature.