UI Changes
Got version 1 of the base UI in place. It is very ugly and will simply
be used to make sure, over time, that everything in the UI is working.
Programming it in a modular sense so that elements can be swapped in and
out, without breaking the communication between the UI and other
entities.