More UI changes.
Changed the UI class (still messy tho) so that UI can be changed easily
by just supplying the type of UI, as long as it is predefined in the
inspector.

Created a version 1 of the province management UI, it will show the
resources that province is producing, the happiness, the population and
the money that province is producing. This can be accessed by clicking
the province point on the province, which will also raise the province
slightly.

This is yet to be implemented dynamically.