Networking
Got networking working how I want it. This is a big milestone for me as
I quit the project so long ago due to not being able to overcome this
issue.

Thus far, getting networking to work with unit movement as a visual
test, the client can tell the server to move the unit and the server
will move the unit which will, subsequently, move the unit for all
clients as if the client itself had moved it.

Got this working with pin placement by having a UnitControl script on
the player object that is instantiated over the network.