Fixed Unit/Client Spawning Issue.
Did some testing, fixed previous issue. Needs cleaning up a little bit
to make more modular.

The players nation is set locally, I choose a nation randomly, but down
the line a nation will actually be chosen. Setting locally avoids the
issue of assigning different nations both serverside and locally,
because all of the setting is done locally. Then on units, I set the
access matrix as an RPC call that check if this units parent nation is
the same as the locally stored client's nation, because it is an RPC
call, if I ask for a local copy of something, it will grab the correct
one dependent upon who the call is happening for, so no longer just the
servers