Tidied up unit access matrix code.
Tidied up the unit access matrix code for use over the network. Sends an
RPC call to all clients, can be called more than once, and should be
called everytime a client connects to make sure that units initially in
the scene send the correct access data to clients connecting after the
fact.