Building System Groundwork
Laid out the groundwork for the building system, got the placeholder UI
to replicate the design I have in mind.

Building object's are created and assigned to inactive and active lists
upon every province's instantiation. Not yet, but in the future, the
buildings in the active list will have their effects occur, or their
actions be able to be triggered whilst the buildings in the inactive
list will not.

When a building object is moved from the inactive to active list or vice
versa, the UI object associated with that building object is moved
up/down depending upon its position in the list, and it is assigned a
new parent which is either the active or inactive panel.

Did some code cleanup, not much.