Basic version 1 Building System implemented
Got the basic functionality for version 1 of the building system in
place.

On each province object, there resides a List of inactive buildings and
a List of active buildings. The Buildings can be moved between the
inactive/active lists and when a province management screen is opened
(clicked on), the province management class will check what lists that
provinces building objects are in, and position them in the correct UI
panel accordingly. This not being an issue as only the UI object is
moved so no effects are changed as the only thing that is the same
between all provinces is the UI object, not whether the object is active
or not, and it also works fine as only one province management screen
can be open at once, and so the positions will always be relative to the
currently active province.