Building system cost and modifiers.
Implemented the cost and modifier factor associated with buildings. When
a building is activated on a province, it will cost resources, if those
resources are available on the province the building will be activated,
and the resources deducted from the province's stock.

When a building is active, that building's modifiers will be added onto
the province's modifiers.

**Currently a bug where when you press the activate building button, the
building that was activated is only being activated on the last listed
province in the heirarchy. Possibly because when setting the parent
province of a building, it goes through all of the provinces, and the
last province was the last one to be set although they should be
separate between provinces.