Building System Finishing Touches and Bug/Issue Fixing
Put the finishing touches on the building system (Temp UI not counting).
It now works on all provinces, buttons to activate the buildings can be
pressed, and will activate the corresponding building and deactivate the
button. The modifiers of a building will the be added onto the
province's modifiers.

Fixed a bunch of issues. The building system previously had the issue
when activating a building and choosing a wrong inactivebuildings index.

When a button was pressed, a number was passed in via the inspector to
represent that building's index in the list of inactive buildings, the
problem was, that building, when activated, got removed from the
inactive list, and so the other buildings indexes change, but I am still
asking for their original index, I tried multiple solutions such as
trying to activate the building at its given index minus the number of
buildings that have been activated. I also figured that if a building is
removed from the inactive list, and then the next building I am trying
to remove's index is below the building that was just removed, its index
stays the same, but then if it is above, it needs to minus 1, that
works, unless you remove 2 buildings, then it will only account for the
very last buildings index, breaking things.

The solution was to give each building an enum, and then when the button
under a building is pressed to activate it, I pass that buildings enum
to my function, the function then goes through the list of inactive
buildings and checks the passed in enum against each building in that
list's enum, when it matches, it calls the currently active province's
activateBuilding function and passes in the Building that matched the
enum. The enums are just called Barracks, Refinery, Fortress

Changed some of the way in which the UI updates values to make it more
readable and maintainable.

Other bugs fixed:
.Fixed where getting a NullReferenceException when the host/client
clicks another province it does not own.
.Made a temporary workaround to fix the problem wherein Start()
functionality is called on local objects before any data has been
initialized, causing errors as no data has been gathered. All start()
functionality gets delayed until the client/host has joined.
.Stopped units from reacting to a click when the province UI was active.