Bug Fixes and misc
Fixed a bug where if you click multiple other province points whilst a
province UI is currently open, the active province will stay raised upon
exiting the UI. Fixed by adding a check within the provinceIsClicked()
to see whether another UI is currently active before actually firing a
raycast.

Fixed another bug where you got a MissingReferenceException for the
infantryUnit when starting a server/client, stopping it, and then
starting it again as the infantry unit was trying to be acced but did
not exist due to the server/client leaving.