Issue fixes and small changes.
Been unmotivated recently, just doing some small stuff whilst I figure
it out.

Changed the setting of a buildings parent position to use SetParent() as
opposed to trans.parent as it caused UI scaling issues.

Made a province's modifiers randomly assigned when a province is created
for future time-span testing.

Added a strength value to the province that is increased by 20 when the
fortress is activated.