Networking re-work #1
Started reworking how I approach networking in the game, more
specifically the storage of data.

When a client joins, it now stores a reference to that client's
DataManager class (which holds all relevant data), it stores this
reference on the server, but then also creates a local copy initially to
allow for UI and other non-networking essential factors to get their
initial values.

At the moment (subject to change), whenever data needs to be used (such
as this client's nation transform), it will be taken from the
DataManager reference that is held on the server. Currently does this by
updating the local reference with the networked reference, and then
returning it when getThisNationTransform() is called via the local
DataManager. This is probably going to change in terms of what is
updated and when it is updated, but the premise will remain the same.